Update RMCLogger.cpp
When a capture folder is created by the plugin, it can be difficult to understand the different game names based on the process name alone. This pull request adds a mapping of process names to more readable game names, so that the capture folders have more understandable names.

The mapping is stored in a std::map object called gameNames, which contains key-value pairs where the key is the process name and the value is the corresponding game name. The code then checks if the process name starts with any of the keys in the gameNames map, and if so, replaces the key with the corresponding value.

For example, if the process name is "ctrapp", the code will replace it with "super-smash-bros3ds". If the process name is "kujira-1", the code will replace it with "Pokemon X". If no match is found, the original process name is used.